Create new util named PasswordHashUtil and add hashingPassword function there
